scrub controller methods
