--- pkg/cmd/server/kubernetes/master/kube_controller_init.go
@@ -1,1 +1,0 @@
-package master
