+++ pkg/cmd/server/origin/controller/quota.go
@@ -2,7 +2,6 @@ package controller
 
 import (
 	"math/rand"
-	"sync"
 	"time"
 
 	"github.com/openshift/origin/pkg/cmd/server/bootstrappolicy"
@@ -47,7 +46,6 @@ func RunResourceQuotaManager(ctx ControllerContext) (bool, error) {
 }
 
 type ClusterQuotaReconciliationControllerConfig struct {
-	Mapper                         clusterquotamapping.ClusterQuotaMapper
 	DefaultResyncPeriod            time.Duration
 	DefaultReplenishmentSyncPeriod time.Duration
 }
@@ -62,9 +60,12 @@ func (c *ClusterQuotaReconciliationControllerConfig) RunController(ctx Controlle
 	)
 	groupKindsToReplenish := quota.AllEvaluatedGroupKinds
 
+	clusterQuotaMappingController := clusterquotamapping.NewClusterQuotaMappingController(
+		ctx.ExternalKubeInformers.Core().V1().Namespaces(),
+		ctx.QuotaInformers.Quota().InternalVersion().ClusterResourceQuotas())
 	options := clusterquotareconciliation.ClusterQuotaReconcilationControllerOptions{
 		ClusterQuotaInformer: ctx.QuotaInformers.Quota().InternalVersion().ClusterResourceQuotas(),
-		ClusterQuotaMapper:   c.Mapper,
+		ClusterQuotaMapper:   clusterQuotaMappingController.GetClusterQuotaMapper(),
 		ClusterQuotaClient:   ctx.ClientBuilder.DeprecatedOpenshiftClientOrDie(saName),
 
 		Registry:     resourceQuotaRegistry,
@@ -77,23 +78,12 @@ func (c *ClusterQuotaReconciliationControllerConfig) RunController(ctx Controlle
 		ReplenishmentResyncPeriod: controller.StaticResyncPeriodFunc(c.DefaultReplenishmentSyncPeriod),
 		GroupKindsToReplenish:     groupKindsToReplenish,
 	}
+	clusterQuotaReconciliationController := clusterquotareconciliation.NewClusterQuotaReconcilationController(options)
+	clusterQuotaMappingController.GetClusterQuotaMapper().AddListener(clusterQuotaReconciliationController)
 
-	controller := clusterquotareconciliation.NewClusterQuotaReconcilationController(options)
-	c.Mapper.AddListener(controller)
-	go controller.Run(5, ctx.Stop)
+	go clusterQuotaMappingController.Run(5, ctx.Stop)
+	go clusterQuotaReconciliationController.Run(5, ctx.Stop)
 
-	return true, nil
-}
-
-type ClusterQuotaMappingControllerConfig struct {
-	ClusterQuotaMappingController *clusterquotamapping.ClusterQuotaMappingController
-}
-
-func (c *ClusterQuotaMappingControllerConfig) RunController(ctx ControllerContext) (bool, error) {
-	var syncOnce sync.Once
-	syncOnce.Do(func() {
-		go c.ClusterQuotaMappingController.Run(5, ctx.Stop)
-	})
 	return true, nil
 }
